refactor(frontend): decouple productivity god class into specialized components and hooks
- Decomposed `productivity.tsx` (858 lines) into manageable specialized units.
- Extracted heavy data filtering and sorting logic into `useProductivityData.ts`.
- Decoupled `ListEmptyComponent` widget tree into `ProductivityEmptyState.tsx`.
- Decoupled inline event mapping UI into `EventCard.tsx`.
- Reduced primary screen file down to 466 lines for extreme readability improvement.
